Move Timeline and Participant to /components
