+++ newdle/client/src/TimelineExample.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Timeline from './Timeline';
+import Timeline from './components/Timeline';
 
 export default function TimelineExample() {
   const availability = [
