+++ newdle/client/src/components/Timeline.module.scss
@@ -1,4 +1,4 @@
-@import 'styles/palette.scss';
+@import '../styles/palette.scss';
 
 $row-height: 50px;
 $label-width: 120px;
